+++ hvega/tests/Test.hs
@@ -46,11 +46,13 @@ import qualified ConfigTests as ConfT
 import qualified DataTests as DT
 import qualified EncodingTests
 import qualified FillStrokeTests as FST
+import qualified FilterTests
 import qualified GeoTests as GT
 import qualified HyperlinkTests as HT
 import qualified ImageTests as ImT
 import qualified InteractionTests as IT
 import qualified LegendTests as LT
+import qualified MarkTests
 import qualified NullTests as NT
 import qualified PositionTests as PT
 import qualified ProjectionTests as PjT
@@ -119,11 +121,13 @@ baseTests = testGroup "base"
   , toTests "Data" "data" DT.testSpecs
   , toTests "Encoding" "encoding" EncodingTests.testSpecs
   , toTests "FillStroke" "fillstroke" FST.testSpecs
+  , toTests "Filter" "filter" FilterTests.testSpecs
   , toTests "Geo" "geo" GT.testSpecs
   , toTests "Hyperlink" "hyperlink" HT.testSpecs
   , toTests "Image" "image" ImT.testSpecs
   , toTests "Interaction" "interaction" IT.testSpecs
   , toTests "Legend" "legend" LT.testSpecs
+  , toTests "Mark" "mark" MarkTests.testSpecs
   , toTests "Null" "null" NT.testSpecs
   , toTests "Position" "position" PT.testSpecs
   , toTests "Projection" "projection" PjT.testSpecs